Allow generator to be used with Chrome 61.
Chrome 61's Protocol definition has a new property on Page.Frame that does not adhere to the existing Protocol schema -- specifically, optional and experimental on the 'unreachableUrl' command are strings. I know!

This commit allows the generator to be used with Chrome 61, bumps dependencies and adds the ability for a pre-defined json schema to be used at the CLI.